requirements.txt
# Transformers & NLP
transformers>=4.36.0
tokenizers>=0.15.0
datasets>=2.15.0
sentence-transformers>=2.2.2

# Training & Optimization
accelerate>=0.25.0
wandb>=0.16.0

# Scientific Computing
numpy<2
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Text Processing & Analysis
nltk>=3.8.1
spacy>=3.7.0
textstat>=0.7.3
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
jupyter>=1.0.0
ipykernel>=6.25.0
einops>=0.7.0

# Evaluation & Metrics
evaluate>=0.4.0
rouge-score>=0.1.2
bert-score>=0.3.13

# Development & Debugging
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0

# Optional: Advanced features (install if needed)
# faiss-cpu>=1.7.4
# umap-learn>=0.5.4
# deepspeed>=0.12.0
# bitsandbytes>=0.41.0